requirements.txt
smbus2==0.4.2
paho-mqtt==1.6.1
python-dotenv==1.0.0
